fix: rename the project
https://github.com/suzuki-shunsuke/fagott/issues/20
