+++ README.md
@@ -1,30 +1,34 @@
-# fagott
+# flute
 
-[![GoDoc](http://img.shields.io/badge/go-documentation-blue.svg?style=flat-square)](http://godoc.org/github.com/suzuki-shunsuke/fagott/fagott)
-[![Build Status](https://cloud.drone.io/api/badges/suzuki-shunsuke/fagott/status.svg)](https://cloud.drone.io/suzuki-shunsuke/fagott)
-[![codecov](https://codecov.io/gh/suzuki-shunsuke/fagott/branch/master/graph/badge.svg)](https://codecov.io/gh/suzuki-shunsuke/fagott)
-[![Go Report Card](https://goreportcard.com/badge/github.com/suzuki-shunsuke/fagott)](https://goreportcard.com/report/github.com/suzuki-shunsuke/fagott)
-[![GitHub last commit](https://img.shields.io/github/last-commit/suzuki-shunsuke/fagott.svg)](https://github.com/suzuki-shunsuke/fagott)
-[![GitHub tag](https://img.shields.io/github/tag/suzuki-shunsuke/fagott.svg)](https://github.com/suzuki-shunsuke/fagott/releases)
-[![License](http://img.shields.io/badge/license-mit-blue.svg?style=flat-square)](https://raw.githubusercontent.com/suzuki-shunsuke/fagott/master/LICENSE)
+[![GoDoc](http://img.shields.io/badge/go-documentation-blue.svg?style=flat-square)](http://godoc.org/github.com/suzuki-shunsuke/flute/flute)
+[![Build Status](https://cloud.drone.io/api/badges/suzuki-shunsuke/flute/status.svg)](https://cloud.drone.io/suzuki-shunsuke/flute)
+[![codecov](https://codecov.io/gh/suzuki-shunsuke/flute/branch/master/graph/badge.svg)](https://codecov.io/gh/suzuki-shunsuke/flute)
+[![Go Report Card](https://goreportcard.com/badge/github.com/suzuki-shunsuke/flute)](https://goreportcard.com/report/github.com/suzuki-shunsuke/flute)
+[![GitHub last commit](https://img.shields.io/github/last-commit/suzuki-shunsuke/flute.svg)](https://github.com/suzuki-shunsuke/flute)
+[![GitHub tag](https://img.shields.io/github/tag/suzuki-shunsuke/flute.svg)](https://github.com/suzuki-shunsuke/flute/releases)
+[![License](http://img.shields.io/badge/license-mit-blue.svg?style=flat-square)](https://raw.githubusercontent.com/suzuki-shunsuke/flute/master/LICENSE)
 
 Golang HTTP Client testing framework
 
 ## Presentation
 
-https://speakerdeck.com/szksh/introduction-of-fagott-golang-http-client-testing-framework
+https://speakerdeck.com/szksh/introduction-of-flute-golang-http-client-testing-framework
+
+## Note: This project name was changed
+
+Please see https://github.com/suzuki-shunsuke/flute/issues/20 .
 
 ## Overview
 
-`fagott` is the Golang HTTP client testing framework.
+`flute` is the Golang HTTP client testing framework.
 The goal is
 
 * Test request parameters such as the request path, headers and body
 * Mock the HTTP server
 
-`*fagott.Transport` implements [http.RoundTripper](https://golang.org/pkg/net/http/#RoundTripper).
+`*flute.Transport` implements [http.RoundTripper](https://golang.org/pkg/net/http/#RoundTripper).
 
-`fagott` uses [testify](https://github.com/stretchr/testify)'s assert internally.
+`flute` uses [testify](https://github.com/stretchr/testify)'s assert internally.
 You can test the http request parameters with assert.
 
 For example, the following test failure message means the request header is unexpected value.
