+++ flute/matcher.go
@@ -1,4 +1,4 @@
-package fagott
+package flute
 
 import (
 	"io/ioutil"
