+++ flute/mock_round_tripper_test.go
@@ -1,4 +1,4 @@
-package fagott
+package flute
 
 // Don't edit this file.
 // This file is generated by gomic 0.5.6.
