+++ flute/response.go
@@ -1,4 +1,4 @@
-package fagott
+package flute
 
 import (
 	"encoding/json"
