+++ flute/transport.go
@@ -1,4 +1,4 @@
-package fagott
+package flute
 
 import (
 	"fmt"
